Added condition check for deleting in controller, hid the edit and delete options if user currently in the session isnt the owner of the submission when viewing in show.ejs

+++ controllers/challenges.js
@@ -74,6 +74,18 @@ const showChallenge = async (req, res)=>{
 }
 
 const deleteChallenge = async (req, res)=>{
+    const foundChallenge = await Challenge.findById(req.params.challengeId)
+
+    if (!foundChallenge){
+        return res.send('Record Does Not Exist.')
+    }
+
+    const isOwner = foundChallenge.owner.equals(req.session.user._id)
+
+    if (!isOwner){
+        return res.send('You do not have permission for this action.')
+    }
+
     await Challenge.findByIdAndDelete(req.params.challengeId)
     res.redirect('/challenges')
 }
